+++ apis/cassandra/v1beta1/cassandradatacenter_types_test.go
@@ -54,3 +54,64 @@ func TestLegacyInternodeDisabled(t *testing.T) {
 
 	assert.False(t, dc.LegacyInternodeEnabled())
 }
+
+func TestUseClientImage(t *testing.T) {
+	assert := assert.New(t)
+
+	tests := []struct {
+		serverType string
+		version    string
+		should     bool
+	}{
+		{
+			serverType: "cassandra",
+			version:    "4.1.0",
+			should:     true,
+		},
+		{
+			serverType: "cassandra",
+			version:    "4.1.2",
+			should:     true,
+		},
+		{
+			serverType: "cassandra",
+			version:    "5.0.0",
+			should:     true,
+		},
+		{
+			serverType: "cassandra",
+			version:    "3.11.17",
+			should:     false,
+		},
+		{
+			serverType: "cassandra",
+			version:    "4.0.8",
+			should:     false,
+		},
+		{
+			serverType: "dse",
+			version:    "6.8.39",
+			should:     false,
+		},
+		{
+			serverType: "dse",
+			version:    "4.1.2",
+			should:     false,
+		},
+	}
+
+	for _, tt := range tests {
+		dc := CassandraDatacenter{
+			Spec: CassandraDatacenterSpec{
+				ServerVersion: tt.version,
+				ServerType:    tt.serverType,
+			},
+		}
+
+		if tt.should {
+			assert.True(dc.UseClientImage())
+		} else {
+			assert.False(dc.UseClientImage())
+		}
+	}
+}