+++ pkg/reconciliation/construct_podtemplatespec.go
@@ -28,6 +28,7 @@ import (
 const (
 	DefaultTerminationGracePeriodSeconds = 120
 	ServerConfigContainerName            = "server-config-init"
+	ServerBaseConfigContainerName        = "server-config-init-base"
 	CassandraContainerName               = "cassandra"
 	PvcName                              = "server-data"
 	SystemLoggerContainerName            = "server-system-logger"
@@ -303,15 +304,28 @@ func addVolumes(dc *api.CassandraDatacenter, baseTemplate *corev1.PodTemplateSpe
 			EmptyDir: &corev1.EmptyDirVolumeSource{},
 		},
 	}
+	volumeDefaults := []corev1.Volume{vServerConfig, vServerLogs}
 
-	vectorLib := corev1.Volume{
-		Name: "vector-lib",
-		VolumeSource: corev1.VolumeSource{
-			EmptyDir: &corev1.EmptyDirVolumeSource{},
-		},
+	if dc.UseClientImage() {
+		vBaseConfig := corev1.Volume{
+			Name: "server-config-base",
+			VolumeSource: corev1.VolumeSource{
+				EmptyDir: &corev1.EmptyDirVolumeSource{},
+			},
+		}
+
+		volumeDefaults = append(volumeDefaults, vBaseConfig)
 	}
 
-	volumeDefaults := []corev1.Volume{vServerConfig, vServerLogs, vectorLib}
+	if !dc.Spec.DisableSystemLoggerSidecar {
+		vectorLib := corev1.Volume{
+			Name: "vector-lib",
+			VolumeSource: corev1.VolumeSource{
+				EmptyDir: &corev1.EmptyDirVolumeSource{},
+			},
+		}
+		volumeDefaults = append(volumeDefaults, vectorLib)
+	}
 
 	if addLegacyInternodeMount {
 		vServerEncryption := corev1.Volume{
@@ -377,11 +391,24 @@ func buildInitContainers(dc *api.CassandraDatacenter, rackName string, baseTempl
 	serverCfg.Name = ServerConfigContainerName
 
 	if serverCfg.Image == "" {
-		if dc.GetConfigBuilderImage() != "" {
-			serverCfg.Image = dc.GetConfigBuilderImage()
+
+		if dc.UseClientImage() {
+			if images.GetClientImage() != "" {
+				serverCfg.Image = images.GetClientImage()
+				serverCfg.Args = []string{
+					"config",
+					"build",
+				}
+			}
 		} else {
-			serverCfg.Image = images.GetConfigBuilderImage()
+			// Use older config-builder
+			if dc.GetConfigBuilderImage() != "" {
+				serverCfg.Image = dc.GetConfigBuilderImage()
+			} else {
+				serverCfg.Image = images.GetConfigBuilderImage()
+			}
 		}
+
 		if images.GetImageConfig() != nil && images.GetImageConfig().ImagePullPolicy != "" {
 			serverCfg.ImagePullPolicy = images.GetImageConfig().ImagePullPolicy
 		}
@@ -392,7 +419,42 @@ func buildInitContainers(dc *api.CassandraDatacenter, rackName string, baseTempl
 		MountPath: "/config",
 	}
 
-	serverCfg.VolumeMounts = combineVolumeMountSlices([]corev1.VolumeMount{serverCfgMount}, serverCfg.VolumeMounts)
+	configMounts := []corev1.VolumeMount{serverCfgMount}
+
+	var configContainer *corev1.Container
+	if dc.UseClientImage() {
+
+		configBaseMount := corev1.VolumeMount{
+			Name:      "server-config-base",
+			MountPath: "/cassandra-base-config",
+		}
+
+		configMounts = append(configMounts, configBaseMount)
+
+		// Similar to k8ssandra 1.x, use config-container if use new config-builder replacement
+		configContainer = &corev1.Container{
+			Name: ServerBaseConfigContainerName,
+		}
+
+		if configContainer.Image == "" {
+			serverImage, err := makeImage(dc)
+			if err != nil {
+				return err
+			}
+
+			configContainer.Image = serverImage
+			if images.GetImageConfig() != nil && images.GetImageConfig().ImagePullPolicy != "" {
+				configContainer.ImagePullPolicy = images.GetImageConfig().ImagePullPolicy
+			}
+
+			configContainer.Command = []string{"/bin/sh"}
+			configContainer.Args = []string{"-c", "cp -rf /etc/cassandra/* /cassandra-base-config/"}
+		}
+
+		configContainer.VolumeMounts = []corev1.VolumeMount{configBaseMount}
+	}
+
+	serverCfg.VolumeMounts = combineVolumeMountSlices(configMounts, serverCfg.VolumeMounts)
 
 	serverCfg.Resources = *getResourcesOrDefault(&dc.Spec.ConfigBuilderResources, &DefaultsConfigInitContainer)
 
@@ -427,7 +489,11 @@ func buildInitContainers(dc *api.CassandraDatacenter, rackName string, baseTempl
 	if !foundOverrides {
 		// Note that append makes a copy, so we must do this after
 		// serverCfg has been properly set up.
-		baseTemplate.Spec.InitContainers = append(baseTemplate.Spec.InitContainers, *serverCfg)
+		if dc.UseClientImage() {
+			baseTemplate.Spec.InitContainers = append(baseTemplate.Spec.InitContainers, *configContainer, *serverCfg)
+		} else {
+			baseTemplate.Spec.InitContainers = append(baseTemplate.Spec.InitContainers, *serverCfg)
+		}
 	}
 
 	return nil