+++ tests/config_change_condition/config_change_condition_suite_test.go
@@ -65,7 +65,7 @@ var _ = Describe(testName, func() {
 			ns.WaitForDatacenterReady(dcName)
 
 			step = "change the config"
-			json := ginkgo_util.CreateTestJson("{\"spec\": {\"config\": {\"cassandra-yaml\": {\"roles_validity_in_ms\": 256000}, \"jvm-server-options\": {\"garbage_collector\": \"CMS\"}}}}")
+			json := ginkgo_util.CreateTestJson("{\"spec\": {\"config\": {\"cassandra-yaml\": {\"roles_validity\": \"256000ms\"}, \"jvm-server-options\": {\"garbage_collector\": \"CMS\"}}}}")
 			k = kubectl.PatchMerge(dcResource, json)
 			ns.ExecAndLog(step, k)